Allow directly searching subtitles.

+++ data/examples/import-subs2srs.js
@@ -1,4 +1,5 @@
 const glob = require("glob");
+process.env["NODE_CONFIG_DIR"] = __dirname + "/../../config/";
 const config = require("config");
 const move = require('move-concurrently');
 const fs = require('fs');
@@ -28,14 +29,13 @@ const models = require("../../models");
 program
   .version('0.0.1')
   .option('-d, --directory [value]', 'Specify the relative directory where the anki file is extracted to (I.e. "subs2srs/HarryPotter/")')
-  .option('-s, --source [value]', 'The title of the source of the anki deck (I.e. Drama Name, Book Name)')
-  .option('-t, --type [value]', 'The type of the source of the anki deck (I.e. Book, Drama, Anime)')
-  .option('-p, --purge', 'If used will purge the database of example sentences')
+  .option('-n, --name [value]', 'The title of the source of the anki deck (I.e. Drama Name, Book Name)')
+  .option('-t, --tag [value]', 'Specify a tag to give the video.')
   .parse(process.argv);
 
 const directory = program.directory;
-const source = program.source;
-const type = program.type;
+const name = program.name;
+const tag = program.tag;
 let entered = 0;
 
 const include = [
@@ -58,19 +58,23 @@ const include = [
     }
 ];
 
+const fileExists = async path => !!(await fs.promises.stat(path).catch(e => false));
+
 (async () => {
     await models.sequelize.sync();
-    if (program.purge) {
-        await models.SentenceExample.destroy({
-            where: {},
-            truncate: true
-        });
-    }
-
     const db = await sqlite.open(path.join(directory, 'collection.anki2'), { Promise })
     console.log('Accessed database.');
     const notes = await db.all('SELECT * FROM notes');
     console.log(`Found ${notes.length} notes.`);
+
+    const exampleMedia = await models.SentenceExampleMedia.create({
+        name: name,
+        type: 'subs2srs',
+        soundPath: '',
+        sourceId: '',
+        tags: tag ? tag.split(',') : []
+    });
+
     let i = 0;
     await Promise.map(notes, async note => {
         const notea = note.flds.split("\x1f");
@@ -83,76 +87,63 @@ const include = [
             return;
         }
 
-        const japanese = notea.find(x => x.match(config.isJapanese) && !x.includes('<') && !x.includes('_'));
-        const english = notea.find(x => !x.match(config.isJapanese) && !x.includes('<') && !x.includes('_') && x.length > 2);
+        const japanese = notea.find(x => x.match(config.isJapanese) && !x.includes('<') && !x.includes('[') && !x.includes('(') && !x.includes('_'));
+        if (!japanese) {
+            return;
+        }
 
         let results = await Promise.fromCallback(callback => mecab.parse(japanese, callback));
         results = results.map(data => {
             return {
-                kanji: data[0],
+                original: data[0],
                 lexical: data[1],
                 compound: data[2],
                 compound2: data[3],
                 compound3: data[4],
                 conjugation: data[5],
                 inflection: data[6],
-                original: data[7],
-                reading: data[8],
-                pronunciation: data[9] || ''
+                root: data[7],
+                pronunciation: data[8],
+                realizedPronunciation: data[9] || ''
             }
         });
 
         let ext = sound.split('.');
         ext = ext[ext.length - 1];
         const oldPath = path.join(directory, 'collection.media', sound);
-        const newSoundFile = `audio/${uuidv1()}.${ext}`;
-        const newSoundPath = `${__dirname}/${newSoundFile}`;
+        if (!(await fileExists(oldPath))) {
+            return;
+        }
+
+        let newSoundFile = `audio/${uuidv1()}.${ext}`;
+        let newSoundPath = `${__dirname}/${newSoundFile}`;
+        while (await fileExists(newSoundPath)) {
+            newSoundFile = `audio/${uuidv1()}.${ext}`;
+            newSoundPath = `${__dirname}/${newSoundFile}`;
+        }
         await fs.promises.copyFile(oldPath, newSoundPath);
 
         const example = await models.SentenceExample.create({
-            japanese: japanese,
-            english: english,
-            soundPath: newSoundFile,
-            type: type,
-            source: source
+            text: japanese,
+            soundPath: newSoundPath
         });
+
+        await exampleMedia.addSentence(example);
 
         await Promise.mapSeries(results, async x => {
             const original = x.original;
-            const reading = x.reading;
-            const readingH = moji(reading || original).convert('KK', 'HG').toString();
-            const jmdict = await models.JMdictEntry.findAll({
-                where: {
-                    [Sequelize.Op.or]: [
-                        {
-                            '$entryKanjiElements.word$': original
-                        },
-                        {
-                            '$entryReadingElements.word$': original
-                        }
-                    ]
-                },
-                include: include
+            const root = x.root;
+            const pronunciation = x.pronunciation;
+
+            const component = await models.SentenceExampleComponent.create({
+                text: original,
+                root: root,
+                pronunciation: pronunciation
             });
 
-            const search = jmdict.filter(x => {
-                const kanjiPart = x.entryKanjiElements && x.entryKanjiElements.some(k => k.word == original);
-                if (kanjiPart) {
-                    return true;
-                }
-
-                return x.entryReadingElements.some(r => r.word == original);
-            });
-
-            if (search.length) {
-                await Promise.mapSeries(search, result => {
-                    entered++;
-                    return result.addSentenceExample(example);
-                });
-            }
-
-            console.log(`%${(100.0 * i) / notes.length } left. ${entered} example sentence/word pairs entered.`);
-            i++;
+            await example.addComponent(component);
         });
-    }, { concurrency: 10 });
+        i++;
+        console.log(`${i} / ${notes.length } completed.`);
+    }, { concurrency: 30 });
 })();
